Key improvements include:   - Enhanced game title normalization :     - Ignores "The " prefix     - Handles leading non-alphanumeric characters. {[`   - Centralized and refactored AlphabetFilter styling:     - Moved theme-specific styles from component CSS to themes.scss.     - Renamed CSS custom properties(`--filter-accent-color` to `--alphabet-filter-accent-color`).   - AlphabetFilter  -> index.tsx
