--- .vscode/easycode.ignore
@@ -1,12 +1,0 @@
-node_modules/
-dist/
-vendor/
-cache/
-.*/
-*.min.*
-*.test.*
-*.spec.*
-*.bundle.*
-*.bundle-min.*
-*.log
-package-lock.json
