+++ src/frontend/screens/Library/components/AlphabetFilter/index.css
@@ -1,37 +1,21 @@
 :root {
-  --filter-accent-color: var(--primary, var(--brand-primary));
-  --filter-accent-hover: var(--primary-hover, var(--brand-primary-hover));
+  --alphabet-filter-accent-color: var(--primary, var(--brand-primary));
+  --alphabet-filter-accent-hover: var(
+    --primary-hover,
+    var(--brand-primary-hover)
+  );
   --filter-text-available: var(--text-secondary);
   --filter-text-hover: var(--text-hover);
   --filter-text-disabled: var(--input-background);
   --filter-bg-hover: var(--filter-bg-hover, rgba(255, 255, 255, 0.12));
-  --filter-shadow-color: rgb(from var(--filter-accent-color) r g b / 0.25);
+  --filter-shadow-color: rgb(
+    from var(--alphabet-filter-accent-color) r g b / 0.25
+  );
 
   --filter-border-radius: 8px;
   --filter-transition-speed: 0.2s ease-in-out;
 }
 
-/* Theme-dynamic accent for active letter */
-body .alphabet-filter-button--active {
-  --filter-accent-color: var(--accent, var(--primary, var(--brand-primary)));
-  --filter-accent-hover: var(
-    --accent-overlay,
-    var(--primary-hover, var(--brand-primary-hover))
-  );
-}
-
-body.zombie .alphabet-filter-button--active,
-body.zombie-classic .alphabet-filter-button--active {
-  --filter-accent-color: var(--primary-button);
-  --filter-accent-hover: var(--primary-button-hover, var(--primary-button));
-}
-
-body.old-school .alphabet-filter-button--active {
-  --filter-accent-color: var(--status-success);
-  --filter-accent-hover: var(--status-success-hover);
-}
-
-/* Alphabet Filter Container */
 .alphabet-filter-container {
   display: flex;
   flex-wrap: wrap;
@@ -40,11 +24,10 @@ body.old-school .alphabet-filter-button--active {
   margin-top: 16px;
 }
 
-/* Alphabet Filter Button Base Styles */
 .alphabet-filter-button {
   background-color: transparent;
   border: none;
-  color: var(--filter-accent-color);
+  color: var(--alphabet-filter-accent-color);
   font-family:
     -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue',
     sans-serif;
@@ -63,7 +46,6 @@ body.old-school .alphabet-filter-button--active {
   outline: none;
 }
 
-/* Alphabet Filter Button States */
 .alphabet-filter-button:not(.alphabet-filter-button--active):not(
     .alphabet-filter-button--disabled
   ):hover {
@@ -73,11 +55,11 @@ body.old-school .alphabet-filter-button--active {
 }
 
 .alphabet-filter-button:focus-visible {
-  box-shadow: 0 0 0 3px var(--filter-accent-color);
+  box-shadow: 0 0 0 3px var(--alphabet-filter-accent-color);
 }
 
 .alphabet-filter-button--active {
-  background-color: var(--filter-accent-color);
+  background-color: var(--alphabet-filter-accent-color);
   color: var(--text-tertiary);
   font-weight: 700;
   box-shadow: 0 4px 12px -2px var(--filter-shadow-color);
@@ -85,7 +67,7 @@ body.old-school .alphabet-filter-button--active {
 }
 
 .alphabet-filter-button--active:hover {
-  background-color: var(--filter-accent-hover);
+  background-color: var(--alphabet-filter-accent-hover);
 }
 
 .alphabet-filter-button--disabled {
